Move embed-exploration to embed-exploration folder
